fix unsigned to negative comparisons

+++ inc/b64.h
@@ -3,7 +3,7 @@
 #pragma once
 
 // B64
-std::string b64_encode(std::string in_string, size_t padded_len = -1);
+std::string b64_encode(std::string in_string, int padded_len = -1);
 // can add a char array => b64 later for images etc
 std::string b64_decode(std::string encoded);
